cleaning up test setup
