+++ packages/.test/polyfillFetch.ts
@@ -1,0 +1,5 @@
+import nodeFetch from 'node-fetch'
+
+// @ts-ignore
+window.fetch = nodeFetch
+
