+++ packages/.test/retry.ts
@@ -1,0 +1,18 @@
+// Like `until` but works off of any assertion, not application code.
+export const retry = (assertion: Function, { interval = 1, timeout = 100 } = {}) => {
+  return new Promise((resolve, reject) => {
+    const startTime = Date.now();
+
+    const tryAgain = () => {
+      setTimeout(() => {
+        try {
+          resolve(assertion());
+        } catch (err) {
+          Date.now() - startTime > timeout ? reject(err) : tryAgain();
+        }
+      }, interval);
+    };
+
+    tryAgain();
+  });
+};
